[feature] Remove Any CPU from pinvoke calls (#57)
* [ignore] Remove Any CPU

* Use 127.0.0.1:0 as TCP local address

* Increased wait time for the GetDiscoveredTopicData call

* Use TCP transport for the TestGetDiscoveredTopicData

* Use other participant to discover topics data

* Use localhost as TCP local address

* Remove LocalAddress from TCP transport

* Assign port to the local/public address

* Use StatusMask::NoStatusMask when removing listener

* Remove debug parameters

* Use TCP transport on the discovery topic tests

* Explicitly clean up entities

* Use default QoS in TestGetDiscoveredTopicData

* Create reader/writer in TestGetDiscoveredTopicData test

* Fix BuiltinTopicKey type (byte[16])

* Fixed platform environment variables

* Adde try-finally blocks

* Added Reader/Writer to the TestGetDiscoveredTopics

* Added PublicAddress to TCP transport

* Removed PublicAddress

* Changed to localhost

* Added parameters to the DCPSInfoRepo process

* Try without transport binding (default)

* Removed WaitForSubscription/Publication checks

* Using UDP transport

* Ignore TopicData test until fix

+++ Sources/OpenDDSharp/BuiltinTopicKey.h
@@ -32,14 +32,14 @@ namespace OpenDDSharp {
 		public value struct BuiltinTopicKey {
 
 		private:
-			array<System::Int32, 1>^ m_value;
+			array<System::Byte, 1>^ m_value;
 
 		public:
 			/// <summary>
 			/// Gets the value of the <see cref="BuiltinTopicKey" />
 			/// </summary>
-			property array<System::Int32, 1>^ Value {
-				array<System::Int32, 1>^ get();
+			property array<System::Byte, 1>^ Value {
+				array<System::Byte, 1>^ get();
 			};
 
 		internal:
